fix(samplingbar): 修复rle_depth类型不匹配的问题
将hw_depth乘以SR_KB(1)的结果强制转换为uint64_t，避免和sw_depth类型不一致导致的编译警告或错误

+++ PXView/pv/toolbars/samplingbar.cpp
@@ -700,7 +700,7 @@ void SamplingBar::update_sample_count_selector() {
   if (mode == LOGIC) {
     _device_agent->get_config_bool(SR_CONF_RLE_SUPPORT, rle_support);
     if (rle_support)
-      rle_depth = min(hw_depth * SR_KB(1), sw_depth);
+      rle_depth = min((uint64_t)(hw_depth * SR_KB(1)), sw_depth);
   }
   // DSO mode (max_timebase/min_timebase) was removed — DSO hardware dropped.
